[refactor] Calculator -> App으로 이름 변경
